requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.6.0
langchain>=0.1.0
langchain-openai>=0.0.5
chromadb>=0.4.22
pydantic>=2.5.0
python-dotenv>=1.0.0
numpy>=1.26.0
pytest>=7.4.0
httpx>=0.25.0
python-multipart>=0.0.6
